backend/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.116.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.1
packaging==25.0
preshed==3.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
shellingham==1.5.4
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
starlette==0.47.1
thinc==8.3.6
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
twilio==8.10.0
python-dotenv==1.0.0
selenium==4.23.1
webdriver-manager==4.0.2
dateparser==1.2.0
